pyproject.toml
[project]
name = "turingmind-arena"
version = "0.1.0"
description = "Multi-agent chat arena with Turing Award laureate personas"
requires-python = ">=3.12"
dependencies = [
    "tornado>=6.4",
    "duckdb>=1.1.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
avatars = ["Pillow>=10.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "handlers"]
